+++ Public/New-ServiceAsync.ps1
@@ -21,26 +21,27 @@ function New-ServiceAsync {
     [CmdletBinding()]
     [OutputType([psobject])]
     param(
-        [Parameter(Mandatory, ValueFromPipeline)]
+        [Parameter(Mandatory, Position = 0, ValueFromPipeline)]
         [ValidateNotNullOrEmpty()]
         [psobject]
         $Space,
 
-        [Parameter(Mandatory)]
+        [Parameter(Mandatory, Position = 1)]
         [ValidateNotNullOrEmpty()]
         [String]
         $ServiceName,
 
-        [Parameter(Mandatory)]
+        [Parameter(Mandatory, Position = 2)]
         [ValidateNotNullOrEmpty()]
         [String]
         $Plan,
 
-        [Parameter(Mandatory)]
+        [Parameter(Mandatory, Position = 3)]
         [ValidateNotNullOrEmpty()]
         [String]
         $Name,
 
+        [Parameter(Position = 4)]
         $params = @()
     )
 
@@ -52,7 +53,7 @@ function New-ServiceAsync {
         Write-Debug "[$($MyInvocation.MyCommand.Name)] PSBoundParameters: $($PSBoundParameters | Out-String)"
         $service = Get-Service -Space $space -Name $servicename
         Write-Debug ($service | ConvertTo-Json -Depth 20)
-        $serviceplans = Get-ServicePlan -Service $service
+        $serviceplans = Get-ServicePlans -Service $service
         Write-Debug ($serviceplans | ConvertTo-Json -Depth 20)
         Write-Output (New-Service $Space $serviceplans $Plan $Name $Params)
     }