+++ cf-api.Tests.ps1
@@ -1,4 +1,0 @@
-#requires -modules Pester
-
-Invoke-Pester -Script @{ Path = 'Public/*' }
-